friendCustomization add content、remove FriendRemark & FriendLabel
