+++ business_meet/build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 group = 'tw.com'
-version = '0.2.7'
+version = '0.2.9'
 sourceCompatibility = '11'
 
 repositories {
