--- business_meet/src/main/java/tw/com/business_meet/bean/FriendLabelBean.java
@@ -1,60 +1,0 @@
-package tw.com.business_meet.bean;
-
-import java.util.Date;
-
-public class FriendLabelBean {
-    private Integer friendLabelNo;
-    private String content;
-    private Integer friendCustomizationNo;
-    private Date createDate;
-    private Date modifyDate;
-    private Integer statusCode;
-
-    public Integer getFriendLabelNo() {
-        return friendLabelNo;
-    }
-
-    public void setFriendLabelNo(Integer friendLabelNo) {
-        this.friendLabelNo = friendLabelNo;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public Integer getFriendCustomizationNo() {
-        return friendCustomizationNo;
-    }
-
-    public void setFriendCustomizationNo(Integer friendCustomizationNo) {
-        this.friendCustomizationNo = friendCustomizationNo;
-    }
-
-    public Date getCreateDate() {
-        return createDate;
-    }
-
-    public void setCreateDate(Date createDate) {
-        this.createDate = createDate;
-    }
-
-    public Date getModifyDate() {
-        return modifyDate;
-    }
-
-    public void setModifyDate(Date modifyDate) {
-        this.modifyDate = modifyDate;
-    }
-
-    public Integer getStatusCode() {
-        return statusCode;
-    }
-
-    public void setStatusCode(Integer statusCode) {
-        this.statusCode = statusCode;
-    }
-}
