--- business_meet/src/main/java/tw/com/business_meet/dao/FriendRemarkDAO.java
@@ -1,10 +1,0 @@
-package tw.com.business_meet.dao;
-
-import tw.com.business_meet.bean.FriendRemarkBean;
-import tw.com.business_meet.vo.FriendRemark;
-
-import java.util.List;
-
-public interface FriendRemarkDAO extends BaseDAO<FriendRemark> {
-    public List<FriendRemark> search(FriendRemarkBean friendRemarkBean);
-}
